remove literal7bit new
