+++ tests/misc.rs
@@ -17,7 +17,7 @@ fn test_loop() -> Result<(), String> {
             AddIf(PC, PC, Nibble::new(2)),  //
             Imm(PC, Literal12Bit::new_checked(14)?), // GOTO: end
             AddImmSigned(A, Literal7Bit::from_signed(-1)?),
-            AddImm(B, Literal7Bit::new(1)),
+            AddImm(B, Literal7Bit::new_checked(1)?),
             Imm(PC, Literal12Bit::new_checked(2)?),
             // LABEL: end=14
             System(Zero, Zero, Nibble::new(SIGHALT)),
